requirements.txt
# V8 Trading Hub Requirements
# Python 3.10+ required (uses match/case, zoneinfo, dataclass features)

# Core
backtrader
pandas>=2.0
numpy>=1.24
matplotlib
streamlit
plotly
requests
python-dateutil
psutil

# Broker APIs
alpaca-py>=0.20

# Market data
yfinance

# Database
asyncpg>=0.29.0
psycopg[binary]>=3.1
psycopg2-binary
sqlalchemy

# Message queue (Redpanda/Kafka IPC for Options process)
confluent-kafka>=2.3

# ML (optional — needed for pop classifier training)
xgboost
scikit-learn
joblib

# Web scraping (Finviz, Fear & Greed)
beautifulsoup4
